[New] `view.ts` and `calcViewPositionAlongAxis` function
